implemented review comments
