+++ binding-library/python/kafka/__init__.py
@@ -1,1 +1,6 @@
 from .kafka import KafkaEvent
+from . import kafka
+ 
+__all__ = (
+ 'KafkaEvent'
+)
