+++ binding-library/python/setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup
 
 setup(
-    name='azure-functions-kafka-trigger',
-    version='4.1',
+    name='azure-functions-kafka-binding',
+    version='1.0',
     packages=['kafka']
 )
